requirements.txt
mysql-connector-python==8.2.0
bcrypt==4.1.0
python-dotenv==1.0.0

# Tkinter é nativo do Python, não precisa instalar!
# Use: python -m tkinter (para testar se funciona)
